+++ cmd/echo.c
@@ -1,6 +1,6 @@
 #include "../minishell.h"
 
-t_ext	g_var;
+extern t_ext	g_var;
 
 int	check_option_n(char *str)
 {
@@ -21,8 +21,8 @@ int	check_option_n(char *str)
 int	vaild_env_name(char c)
 {
 	if (!('a' <= c && c <= 'z')
-			&& !('A' <= c && c <= 'Z')
-			&& !(c == '_'))
+		&& !('A' <= c && c <= 'Z')
+		&& !(c == '_'))
 		return (0);
 	return (1);
 }
