+++ cmd/exit.c
@@ -14,13 +14,13 @@ int	is_num(char *str)
 	return (SUCCESS);
 }
 
-int	end_shell(int flag)
+int	end_shell(int flag, int exit_status)
 {
 	if (flag != 1)
 	{
 		printf(COLOR_RESET);
 		tcsetattr(STDIN_FILENO, TCSANOW, &g_var.restore);
-		exit(0);
+		exit(exit_status);
 	}
 	return (0);
 }
@@ -32,7 +32,7 @@ int	ft_exit(char **cmd, int flag)
 	exit_status = 0;
 	printf("exit\n");
 	if (cmd[1] == NULL)
-		end_shell(flag);
+		end_shell(flag, exit_status);
 	else if (cmd[1] && cmd[2])
 	{
 		exit_status = 1;
@@ -42,12 +42,12 @@ int	ft_exit(char **cmd, int flag)
 	{
 		exit_status = 255;
 		printf("minishell: numeric argument required\n");
-		end_shell(flag);
+		end_shell(flag, exit_status);
 	}
 	else if (cmd[1])
 	{
 		exit_status = ft_atoi(cmd[1]);
-		end_shell(flag);
+		end_shell(flag, exit_status);
 	}
 	return (exit_status);
 }
