+++ cmd/unset.c
@@ -2,13 +2,8 @@
 
 extern t_ext	g_var;
 
-int	ft_unset(t_env **env, char **cmd)
+int	ft_unset2(t_env **env, char **cmd, char *find)
 {
-	t_env	*tmp;
-	char	*find;
-
-	find = cmd[1];
-
 	if (vaild_env_name(cmd[1][0]) == 0)
 	{
 		printf("minichell: %s: not a valid identifier.\n", cmd[1]);
@@ -21,6 +16,19 @@ int	ft_unset(t_env **env, char **cmd)
 		free((*env));
 		return (SUCCESS);
 	}
+	return (2);
+}
+
+int	ft_unset(t_env **env, char **cmd)
+{
+	t_env	*tmp;
+	char	*find;
+	int		result;
+
+	find = cmd[1];
+	result = ft_unset2(env, cmd, find);
+	if (result != 2)
+		return (result);
 	while ((*env) && (*env)->next)
 	{
 		if (ft_strcmp((*env)->next->name, find) == 0)
