+++ minishell.h
@@ -103,6 +103,7 @@ int		ft_semicolon(t_cmd *c);
 int		vaild_env_name(char c);
 int		check_open(int in, t_cmd *c);
 int		exec_cmd(t_cmd *c);
+void	not_builtin(t_cmd *c);
 int		use_builtin(t_cmd *c, t_env *e);
 int		use_redirect(t_cmd *c);
 int		run_cmd(char **cmd, t_env *env);
