+++ signal.c
@@ -21,7 +21,7 @@ void	sigint_handler(int signo)
 	}
 	write(STDOUT_FILENO, "\n", 1);
 	rl_on_new_line();
-	// rl_replace_line("", 0);
+	rl_replace_line("", 0);
 	rl_redisplay();
 }
